Fix: Replaced `[inaudible]` with `instance`
Although proper English would be “uses an instance of”, just replaced it with “uses instance of” because that's what it sounds like was spoken.  This presenter (Aymeric Bard) doesn't have the best English, and the existing subtitles seem to reflect what he literally said, even if grammatical articles are missing here and there.